refactor(headers): split mint_melt.ts into mint_header.ts + melt_header.ts + shared entry helpers
Every other header in src/headers is one class per file named after the
header (fee.ts, shielded_outputs.ts, unshield_balance.ts, melt_header.ts).
mint_melt.ts broke that convention: it held the shared entry helpers AND
the MintHeader class, so MintHeader was the only header with no eponymous
file while MeltHeader already had melt_header.ts — an asymmetry.

Restore symmetry:
- new src/headers/mint_melt_entry.ts holds the shared building blocks
  (IMintMeltEntry, MAX_MINT_MELT_ENTRIES, ENTRY_SIZE, validateMintMeltEntry,
  validateMintMeltEntries, serializeMintMeltEntries,
  deserializeMintMeltEntries).
- new src/headers/mint_header.ts holds MintHeader (imports the helpers).
- melt_header.ts now imports from mint_melt_entry.
- delete mint_melt.ts.
- index.ts / parser.ts import MintHeader from ./mint_header.

Free helper functions (not a TS base class) are the idiomatic equivalent
of hathor-core's _MintMeltHeaderBase here, and keep each header explicit.
Wire format is byte-identical — this is a pure file reorganization.

Tests reshuffled to match: mint_melt_entry.test.ts (the helper functions)
and mint_header.test.ts (MintHeader) replace mint_melt.test.ts;
melt_header.test.ts / parser.test.ts import paths updated. All 127 header
tests pass; lint and tsc clean.
