+++ __tests__/headers/mint_header.test.ts
@@ -5,26 +5,26 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { MeltHeader } from '../../src/headers/melt_header';
-import { MAX_MINT_MELT_ENTRIES, IMintMeltEntry } from '../../src/headers/mint_melt_entry';
+import { MintHeader } from '../../src/headers/mint_header';
+import { IMintMeltEntry, MAX_MINT_MELT_ENTRIES } from '../../src/headers/mint_melt_entry';
 import { VertexHeaderId, getVertexHeaderIdBuffer } from '../../src/headers/types';
 import Network from '../../src/models/network';
 
 const network = new Network('testnet');
 
-describe('MeltHeader', () => {
+describe('MintHeader', () => {
   describe('constructor', () => {
     it('stores the entries verbatim when valid', () => {
       const entries: IMintMeltEntry[] = [
         { tokenIndex: 1, amount: 10n },
         { tokenIndex: 2, amount: 20n },
       ];
-      const header = new MeltHeader(entries);
+      const header = new MintHeader(entries);
       expect(header.entries).toEqual(entries);
     });
 
     it('rejects an empty entries list', () => {
-      expect(() => new MeltHeader([])).toThrow(/MeltHeader requires at least 1 entry/);
+      expect(() => new MintHeader([])).toThrow(/MintHeader requires at least 1 entry/);
     });
 
     it('rejects more than MAX_MINT_MELT_ENTRIES entries (count check fires before the per-entry loop)', () => {
@@ -35,22 +35,22 @@ describe('MeltHeader', () => {
           amount: 1n,
         })
       );
-      expect(() => new MeltHeader(entries)).toThrow(
-        /MeltHeader: too many entries: 17 exceeds maximum 16/
+      expect(() => new MintHeader(entries)).toThrow(
+        /MintHeader: too many entries: 17 exceeds maximum 16/
       );
     });
 
     it('rejects duplicate tokenIndex values', () => {
       const entries: IMintMeltEntry[] = [
-        { tokenIndex: 3, amount: 1n },
-        { tokenIndex: 3, amount: 1n },
+        { tokenIndex: 2, amount: 1n },
+        { tokenIndex: 2, amount: 1n },
       ];
-      expect(() => new MeltHeader(entries)).toThrow(/MeltHeader: duplicate token_index 3/);
+      expect(() => new MintHeader(entries)).toThrow(/MintHeader: duplicate token_index 2/);
     });
 
     it('rejects an entry that fails per-entry validation', () => {
-      expect(() => new MeltHeader([{ tokenIndex: 0, amount: 1n }])).toThrow(/MeltHeader:/);
-      expect(() => new MeltHeader([{ tokenIndex: 1, amount: 0n }])).toThrow(/MeltHeader:/);
+      const entries: IMintMeltEntry[] = [{ tokenIndex: 0, amount: 1n }];
+      expect(() => new MintHeader(entries)).toThrow(/MintHeader:/);
     });
   });
 
@@ -60,16 +60,16 @@ describe('MeltHeader', () => {
       { tokenIndex: 3, amount: 1_000_000n },
     ];
 
-    it('serializeFields prepends the MeltHeader id byte', () => {
-      const header = new MeltHeader(entries);
+    it('serializeFields prepends the MintHeader id byte', () => {
+      const header = new MintHeader(entries);
       const acc: Buffer[] = [];
       header.serializeFields(acc);
       expect(acc.length).toBe(2);
-      expect(acc[0].equals(getVertexHeaderIdBuffer(VertexHeaderId.MELT_HEADER))).toBe(true);
+      expect(acc[0].equals(getVertexHeaderIdBuffer(VertexHeaderId.MINT_HEADER))).toBe(true);
     });
 
     it('serialize, serializeFields and serializeSighash yield identical bytes', () => {
-      const header = new MeltHeader(entries);
+      const header = new MintHeader(entries);
       const a: Buffer[] = [];
       const b: Buffer[] = [];
       const c: Buffer[] = [];
@@ -82,28 +82,28 @@ describe('MeltHeader', () => {
 
     describe('post-construction mutation is caught at serialize time', () => {
       it('rejects a duplicate tokenIndex pushed after construction', () => {
-        const header = new MeltHeader([{ tokenIndex: 1, amount: 10n }]);
+        const header = new MintHeader([{ tokenIndex: 1, amount: 10n }]);
         header.entries.push({ tokenIndex: 1, amount: 20n });
-        expect(() => header.serialize([])).toThrow(/MeltHeader: duplicate token_index 1/);
+        expect(() => header.serialize([])).toThrow(/MintHeader: duplicate token_index 1/);
       });
 
       it('rejects an amount mutated to 0 after construction', () => {
-        const header = new MeltHeader([{ tokenIndex: 1, amount: 10n }]);
+        const header = new MintHeader([{ tokenIndex: 1, amount: 10n }]);
         header.entries[0].amount = 0n;
-        expect(() => header.serialize([])).toThrow(/MeltHeader: amount must be in/);
+        expect(() => header.serialize([])).toThrow(/MintHeader: amount must be in/);
       });
 
       it('rejects an entry array emptied after construction', () => {
-        const header = new MeltHeader([{ tokenIndex: 1, amount: 10n }]);
+        const header = new MintHeader([{ tokenIndex: 1, amount: 10n }]);
         header.entries.length = 0;
-        expect(() => header.serialize([])).toThrow(/MeltHeader requires at least 1 entry/);
+        expect(() => header.serialize([])).toThrow(/MintHeader requires at least 1 entry/);
       });
 
       it('catches the same drift via serializeFields and serializeSighash', () => {
-        const header = new MeltHeader([{ tokenIndex: 1, amount: 10n }]);
+        const header = new MintHeader([{ tokenIndex: 1, amount: 10n }]);
         header.entries[0].tokenIndex = 0;
-        expect(() => header.serializeFields([])).toThrow(/MeltHeader:/);
-        expect(() => header.serializeSighash([])).toThrow(/MeltHeader:/);
+        expect(() => header.serializeFields([])).toThrow(/MintHeader:/);
+        expect(() => header.serializeSighash([])).toThrow(/MintHeader:/);
       });
     });
   });
@@ -114,34 +114,34 @@ describe('MeltHeader', () => {
         { tokenIndex: 1, amount: 10n },
         { tokenIndex: 4, amount: 2n ** 50n },
       ];
-      const original = new MeltHeader(entries);
+      const original = new MintHeader(entries);
       const acc: Buffer[] = [];
       original.serialize(acc);
       const wire = Buffer.concat(acc);
 
-      const [parsed, leftover] = MeltHeader.deserialize(wire, network);
-      expect(parsed).toBeInstanceOf(MeltHeader);
-      expect((parsed as MeltHeader).entries).toEqual(entries);
+      const [parsed, leftover] = MintHeader.deserialize(wire, network);
+      expect(parsed).toBeInstanceOf(MintHeader);
+      expect((parsed as MintHeader).entries).toEqual(entries);
       expect(leftover.length).toBe(0);
     });
 
     it('preserves trailing bytes from the source buffer', () => {
       const entries: IMintMeltEntry[] = [{ tokenIndex: 2, amount: 7n }];
-      const header = new MeltHeader(entries);
+      const header = new MintHeader(entries);
       const acc: Buffer[] = [];
       header.serialize(acc);
       const trailing = Buffer.from([0xaa, 0xbb]);
       const wire = Buffer.concat([Buffer.concat(acc), trailing]);
 
-      const [, leftover] = MeltHeader.deserialize(wire, network);
+      const [, leftover] = MintHeader.deserialize(wire, network);
       expect(leftover.equals(trailing)).toBe(true);
     });
 
-    it('rejects a buffer that does not start with the MeltHeader id', () => {
-      // 0x14 is MintHeader's id — wrong for MeltHeader.deserialize
-      const wrongId = Buffer.from([0x14, 0x01, 0x01, 0, 0, 0, 0, 0, 0, 0, 1]);
-      expect(() => MeltHeader.deserialize(wrongId, network)).toThrow(
-        /Invalid vertex header id for melt header/
+    it('rejects a buffer that does not start with the MintHeader id', () => {
+      // 0x15 is MeltHeader's id — wrong for MintHeader.deserialize
+      const wrongId = Buffer.from([0x15, 0x01, 0x01, 0, 0, 0, 0, 0, 0, 0, 1]);
+      expect(() => MintHeader.deserialize(wrongId, network)).toThrow(
+        /Invalid vertex header id for mint header/
       );
     });
   });
