+++ src/headers/index.ts
@@ -9,7 +9,7 @@ export { default as Header } from './base';
 export { default as FeeHeader } from './fee';
 export { default as ShieldedOutputsHeader } from './shielded_outputs';
 export { default as UnshieldBalanceHeader } from './unshield_balance';
-export { MintHeader } from './mint_melt';
+export { MintHeader } from './mint_header';
 export { MeltHeader } from './melt_header';
 export * from './parser';
 export * from './types';
